refactor: split 'ui.py' into multiple 'ui_*.py' files
