+++ vunit/ui.py
@@ -264,18 +264,16 @@ import traceback
 import logging
 import json
 import os
-from os.path import exists, abspath, join, basename, splitext, normpath, dirname
-from glob import glob
+from os.path import exists, abspath, join, basename, normpath, dirname
 from fnmatch import fnmatch
 from vunit.database import PickledDataBase, DataBase
 from vunit import ostools
 from vunit.vunit_cli import VUnitCLI
 from vunit.simulator_factory import SIMULATOR_FACTORY
-from vunit.simulator_interface import is_string_not_iterable, SimulatorInterface
+from vunit.simulator_interface import SimulatorInterface
 from vunit.color_printer import COLOR_PRINTER, NO_COLOR_PRINTER
 
 from vunit.project import Project
-from vunit.source_file import file_type_of, FILE_TYPES, VERILOG_FILE_TYPES
 from vunit.test_runner import TestRunner
 from vunit.test_report import TestReport
 from vunit.test_bench_list import TestBenchList
@@ -283,9 +281,17 @@ from vunit.exceptions import CompileError
 from vunit.location_preprocessor import LocationPreprocessor
 from vunit.check_preprocessor import CheckPreprocessor
 from vunit.parsing.encodings import HDL_FILE_ENCODING
-from vunit.builtins import Builtins, add_verilog_include_dir
+from vunit.builtins import Builtins
 from vunit.vhdl_standard import VHDL
-from vunit.com import codec_generator
+
+from vunit.ui_common import select_vhdl_standard, check_not_empty
+from vunit.ui_source import SourceFile, SourceFileList
+from vunit.ui_test import Test  # pylint: disable=unused-import
+from vunit.ui_testbench import TestBench  # pylint: disable=unused-import
+from vunit.ui_packagefacade import PackageFacade  # pylint: disable=unused-import
+from vunit.ui_library import Library
+from vunit.ui_results import Results
+
 
 LOGGER = logging.getLogger(__name__)
 
@@ -1283,944 +1289,3 @@ avoid location preprocessing of other functions sharing name with a VUnit log or
                 for source_file in source_files
             ]
         )
-
-
-class Library(object):
-    """
-    User interface of a library
-    """
-
-    def __init__(self, library_name, parent, project, test_bench_list):
-        self._library_name = library_name
-        self._parent = parent
-        self._project = project
-        self._test_bench_list = test_bench_list
-
-    @property
-    def name(self):
-        """
-        The name of the library
-        """
-        return self._library_name
-
-    def set_generic(self, name, value, allow_empty=False):
-        """
-        Set a value of generic within all |configurations| of test benches and tests this library
-
-        :param name: The name of the generic
-        :param value: The value of the generic
-        :param allow_empty: To disable an error when no test benches were found
-
-        :example:
-
-        .. code-block:: python
-
-           lib.set_generic("data_width", 16)
-
-        .. note::
-           Only affects test benches added *before* the generic is set.
-        """
-        for test_bench in self.get_test_benches(allow_empty=allow_empty):
-            test_bench.set_generic(name.lower(), value)
-
-    def set_parameter(self, name, value, allow_empty=False):
-        """
-        Set a value of parameter within all |configurations| of test benches and tests this library
-
-        :param name: The name of the parameter
-        :param value: The value of the parameter
-        :param allow_empty: To disable an error when no test benches were found
-
-        :example:
-
-        .. code-block:: python
-
-           lib.set_parameter("data_width", 16)
-
-        .. note::
-           Only affects test benches added *before* the parameter is set.
-        """
-        for test_bench in self.get_test_benches(allow_empty=allow_empty):
-            test_bench.set_generic(name, value)
-
-    def set_sim_option(self, name, value, allow_empty=False, overwrite=True):
-        """
-        Set simulation option within all |configurations| of test benches and tests this library
-
-        :param name: |simulation_options|
-        :param value: The value of the simulation option
-        :param allow_empty: To disable an error when no test benches were found
-        :param overwrite: To overwrite the option or append to the existing value
-
-        :example:
-
-        .. code-block:: python
-
-           lib.set_sim_option("ghdl.flags", ["--no-vital-checks"])
-
-        .. note::
-           Only affects test benches added *before* the option is set.
-        """
-        for test_bench in self.get_test_benches(allow_empty=allow_empty):
-            test_bench.set_sim_option(name, value, overwrite)
-
-    def set_compile_option(self, name, value, allow_empty=False):
-        """
-        Set compile option for all files within the library
-
-        :param name: |compile_option|
-        :param value: The value of the compile option
-        :param allow_empty: To disable an error when no source files were found
-
-        :example:
-
-        .. code-block:: python
-
-           lib.set_compile_option("ghdl.flags", ["--no-vital-checks"])
-
-
-        .. note::
-           Only affects files added *before* the option is set.
-        """
-        for source_file in self.get_source_files(allow_empty=allow_empty):
-            source_file.set_compile_option(name, value)
-
-    def add_compile_option(self, name, value, allow_empty=False):
-        """
-        Add compile option to all files within the library
-
-        :param name: |compile_option|
-        :param value: The value of the compile option
-        :param allow_empty: To disable an error when no source files were found
-
-        .. note::
-           Only affects files added *before* the option is set.
-        """
-        for source_file in self.get_source_files(allow_empty=allow_empty):
-            source_file.add_compile_option(name, value)
-
-    def get_source_file(self, file_name):
-        """
-        Get a source file within this library
-
-        :param file_name: The name of the file as a relative or absolute path
-
-        :returns: A :class:`.SourceFile` object
-        """
-        return self._parent.get_source_file(file_name, self._library_name)
-
-    def get_source_files(self, pattern="*", allow_empty=False):
-        """
-        Get a list of source files within this libary
-
-        :param pattern: A wildcard pattern matching either an absolute or relative path
-        :param allow_empty: To disable an error if no files matched the pattern
-        :returns: A :class:`.SourceFileList` object
-        """
-        return self._parent.get_source_files(pattern, self._library_name, allow_empty)
-
-    def add_source_files(  # pylint: disable=too-many-arguments
-        self,
-        pattern,
-        preprocessors=None,
-        include_dirs=None,
-        defines=None,
-        allow_empty=False,
-        vhdl_standard=None,
-        no_parse=False,
-        file_type=None,
-    ):
-        """
-        Add source files matching wildcard pattern to library
-
-        :param pattern: A wildcard pattern match the files to add
-        :param include_dirs: A list of include directories
-        :param defines: A dictionary containing Verilog defines to be set
-        :param allow_empty: To disable an error if no files matched the pattern
-        :param vhdl_standard: The VHDL standard used to compile files, if None library default is used
-        :param no_parse: Do not parse file(s) for dependency or test scanning purposes
-        :param file_type: The type of the file; ``"vhdl"``, ``"verilog"``  or ``"systemverilog"``.
-                          Auto-detected by default when set to ``None``.
-        :returns: A list of files (:class:`.SourceFileList`) which were added
-
-        :example:
-
-        .. code-block:: python
-
-           library.add_source_files("*.vhd")
-
-        """
-        if is_string_not_iterable(pattern):
-            patterns = [pattern]
-        else:
-            patterns = pattern
-
-        file_names = []
-        for pattern_instance in patterns:
-            new_file_names = glob(pattern_instance)
-            check_not_empty(
-                new_file_names,
-                allow_empty,
-                "Pattern %r did not match any file" % pattern_instance,
-            )
-            file_names += new_file_names
-
-        return SourceFileList(
-            source_files=[
-                self.add_source_file(
-                    file_name,
-                    preprocessors,
-                    include_dirs,
-                    defines,
-                    vhdl_standard,
-                    no_parse=no_parse,
-                    file_type=file_type,
-                )
-                for file_name in file_names
-            ]
-        )
-
-    def add_source_file(  # pylint: disable=too-many-arguments
-        self,
-        file_name,
-        preprocessors=None,
-        include_dirs=None,
-        defines=None,
-        vhdl_standard=None,
-        no_parse=False,
-        file_type=None,
-    ):
-        """
-        Add source file to library
-
-        :param file_name: The name of the file
-        :param include_dirs: A list of include directories
-        :param defines: A dictionary containing Verilog defines to be set
-        :param vhdl_standard: The VHDL standard used to compile this file, if None library default is used
-        :param no_parse: Do not parse file for dependency or test scanning purposes
-        :param file_type: The type of the file; ``"vhdl"``, ``"verilog"``  or ``"systemverilog"``.
-                          Auto-detected by default when set to ``None``.
-        :returns: The :class:`.SourceFile` which was added
-
-        :example:
-
-        .. code-block:: python
-
-           library.add_source_file("file.vhd")
-
-        """
-        file_name = abspath(file_name)
-
-        if file_type is None:
-            file_type = file_type_of(file_name)
-        elif file_type not in FILE_TYPES:
-            raise ValueError("file_type %r not in %r" % (file_type, FILE_TYPES))
-
-        if file_type in VERILOG_FILE_TYPES:
-            include_dirs = include_dirs if include_dirs is not None else []
-            include_dirs = add_verilog_include_dir(include_dirs)
-
-        new_file_name = self._parent._preprocess(  # pylint: disable=protected-access
-            self._library_name, file_name, preprocessors
-        )
-
-        source_file = self._project.add_source_file(
-            new_file_name,
-            self._library_name,
-            file_type=file_type,
-            include_dirs=include_dirs,
-            defines=defines,
-            vhdl_standard=vhdl_standard,
-            no_parse=no_parse,
-        )
-        # To get correct tb_path generic
-        source_file.original_name = file_name
-
-        self._test_bench_list.add_from_source_file(source_file)
-
-        return SourceFile(source_file, self._project, self._parent)
-
-    def package(self, name):
-        """
-        Get a package within the library
-        """
-        library = self._project.get_library(self._library_name)
-        design_unit = library.primary_design_units.get(name)
-
-        if design_unit is None:
-            raise KeyError(name)
-        if design_unit.unit_type != "package":
-            raise KeyError(name)
-
-        return PackageFacade(self._parent, self._library_name, name, design_unit)
-
-    def entity(self, name):
-        """
-        Get an entity within the library
-
-        :param name: The name of the entity
-        :returns: A :class:`.TestBench` object
-        :raises: KeyError
-        """
-        name = name.lower()
-        library = self._project.get_library(self._library_name)
-        if not library.has_entity(name):
-            raise KeyError(name)
-
-        return self.test_bench(name)
-
-    def module(self, name):
-        """
-        Get a module within the library
-
-        :param name: The name of the module
-        :returns: A :class:`.TestBench` object
-        :raises: KeyError
-        """
-        library = self._project.get_library(self._library_name)
-        if name not in library.modules:
-            raise KeyError(name)
-
-        return self.test_bench(name)
-
-    def test_bench(self, name):
-        """
-        Get a test bench within this library
-
-        :param name: The name of the test bench
-        :returns: A :class:`.TestBench` object
-        :raises: KeyError
-        """
-        name = name.lower()
-
-        return TestBench(
-            self._test_bench_list.get_test_bench(self._library_name, name), self
-        )
-
-    def get_test_benches(self, pattern="*", allow_empty=False):
-        """
-        Get a list of test benches
-
-        :param pattern: A wildcard pattern matching the test_bench name
-        :param allow_empty: To disable an error when no test benches were found
-        :returns: A list of :class:`.TestBench` objects
-        """
-        results = []
-        for test_bench in self._test_bench_list.get_test_benches_in_library(
-            self._library_name
-        ):
-            if not fnmatch(abspath(test_bench.name), pattern):
-                continue
-
-            results.append(TestBench(test_bench, self))
-
-        return check_not_empty(
-            results,
-            allow_empty,
-            "No test benches found within library %s" % self._library_name,
-        )
-
-
-class TestBench(object):
-    """
-    User interface of a test bench.
-    A test bench consists of one or more :class:`.Test` cases. Setting options for a test
-    bench will apply that option all test cases belonging to that test bench.
-    """
-
-    def __init__(self, test_bench, library):
-        self._test_bench = test_bench
-        self._library = library
-
-    @property
-    def name(self):
-        """
-        :returns: The entity or module name of the test bench
-        """
-        return self._test_bench.name
-
-    @property
-    def library(self):
-        """
-        :returns: The library that contains this test bench
-        """
-        return self._library
-
-    def set_attribute(self, name, value):
-        """
-        Set a value of attribute within all |configurations| of this test bench or test cases within it
-
-        :param name: The name of the attribute
-        :param value: The value of the atrribute
-
-        :example:
-
-        .. code-block:: python
-
-           test_bench.set_attribute(".foo", "bar")
-
-        """
-        self._test_bench.set_attribute(name, value)
-
-    def set_generic(self, name, value):
-        """
-        Set a value of generic within all |configurations| of this test bench or test cases within it
-
-        :param name: The name of the generic
-        :param value: The value of the generic
-
-        :example:
-
-        .. code-block:: python
-
-           test_bench.set_generic("data_width", 16)
-
-        """
-        self._test_bench.set_generic(name.lower(), value)
-
-    def set_parameter(self, name, value):
-        """
-        Set a value of parameter within all |configurations| of this test bench or test cases within it
-
-        :param name: The name of the parameter
-        :param value: The value of the parameter
-
-        :example:
-
-        .. code-block:: python
-
-           test_bench.set_parameter("data_width", 16)
-
-        """
-        self._test_bench.set_generic(name, value)
-
-    def set_sim_option(self, name, value, overwrite=True):
-        """
-        Set simulation option within all |configurations| of this test bench or test cases within it
-
-        :param name: |simulation_options|
-        :param value: The value of the simulation option
-        :param overwrite: To overwrite the option or append to the existing value
-
-        :example:
-
-        .. code-block:: python
-
-           test_bench.set_sim_option("ghdl.flags", ["--no-vital-checks"])
-
-        """
-        self._test_bench.set_sim_option(name, value, overwrite)
-
-    def set_pre_config(self, value):
-        """
-        Set :ref:`pre_config <pre_and_post_hooks>` function of all
-        |configurations| of this test bench or test cases within it
-
-        :param value: The pre_config function
-        """
-        self._test_bench.set_pre_config(value)
-
-    def set_post_check(self, value):
-        """
-        Set :ref:`post_check <pre_and_post_hooks>` function of all
-        |configurations| of this test bench or test cases within it
-
-        :param value: The post_check function
-        """
-        self._test_bench.set_post_check(value)
-
-    def add_config(  # pylint: disable=too-many-arguments
-        self,
-        name,
-        generics=None,
-        parameters=None,
-        pre_config=None,
-        post_check=None,
-        sim_options=None,
-        attributes=None,
-    ):
-        """
-        Add a configuration of this test bench or to all test cases within it by copying the default configuration.
-
-        Multiple configuration may be added one after another.
-        If no |configurations| are added the default configuration is used.
-
-        :param name: The name of the configuration. Will be added as a suffix on the test name
-        :param generics: A `dict` containing the generics to be set in addition to the default configuration
-        :param parameters: A `dict` containing the parameters to be set in addition to the default configuration
-        :param pre_config: A :ref:`callback function <pre_and_post_hooks>` to be called before test execution.
-        :param post_check: A :ref:`callback function <pre_and_post_hooks>` to be called after test execution.
-        :param sim_options: A `dict` containing the sim_options to be set in addition to the default configuration
-        :param attributes: A `dict` containing the attributes to be set in addition to the default configuration
-
-        :example:
-
-        Given a test bench that by default gives rise to the test
-        ``lib.test_bench`` and the following ``add_config`` calls:
-
-        .. code-block:: python
-
-           for data_width in range(14, 15+1):
-               for sign in [False, True]:
-                   test_bench.add_config(
-                       name="data_width=%s,sign=%s" % (data_width, sign),
-                       generics=dict(data_width=data_width, sign=sign))
-
-        The following tests will be created:
-
-        * ``lib.test_bench.data_width=14,sign=False``
-
-        * ``lib.test_bench.data_width=14,sign=True``
-
-        * ``lib.test_bench.data_width=15,sign=False``
-
-        * ``lib.test_bench.data_width=15,sign=True``
-
-        """
-        generics = {} if generics is None else generics
-        generics = lower_generics(generics)
-        parameters = {} if parameters is None else parameters
-        generics.update(parameters)
-        attributes = {} if attributes is None else attributes
-        self._test_bench.add_config(
-            name=name,
-            generics=generics,
-            pre_config=pre_config,
-            post_check=post_check,
-            sim_options=sim_options,
-            attributes=attributes,
-        )
-
-    def test(self, name):
-        """
-        Get a test within this test bench
-
-        :param name: The name of the test
-        :returns: A :class:`.Test` object
-        """
-        return Test(self._test_bench.get_test_case(name))
-
-    def get_tests(self, pattern="*"):
-        """
-        Get a list of tests
-
-        :param pattern: A wildcard pattern matching the test name
-        :returns: A list of :class:`.Test` objects
-        """
-        results = []
-        for test_case in self._test_bench.tests:
-            if not fnmatch(test_case.name, pattern):
-                continue
-
-            results.append(Test(test_case))
-        return results
-
-    def scan_tests_from_file(self, file_name):
-        """
-        Scan tests from another file than the one containg the test
-        bench.  Useful for when the top level test bench does not
-        contain the tests.
-
-        Such a structure is not the preferred way of doing things in
-        VUnit but this method exists to accommodate legacy needs.
-
-        :param file_name: The name of another file to scan for tests
-
-        .. warning::
-           The nested module containing the tests needs to be given
-           the ``runner_cfg`` parameter or generic by the
-           instantiating top level test bench. The nested module
-           should not call its parameter or generic `runner_cfg` but
-           rather `nested_runner_cfg` to avoid the VUnit test scanner
-           detecting and running it as a test bench. In SystemVerilog
-           the ``NESTED_TEST_SUITE`` macro should be used instead of
-           the ``TEST_SUITE`` macro.
-        """
-        self._test_bench.scan_tests_from_file(file_name)
-
-
-class PackageFacade(object):
-    """
-    User interface of a Package
-    """
-
-    def __init__(self, parent, library_name, package_name, design_unit):
-        self._parent = parent
-        self._library_name = library_name
-        self._package_name = package_name
-        self._design_unit = design_unit
-
-    def generate_codecs(
-        self, codec_package_name=None, used_packages=None, output_file_name=None
-    ):
-        """
-        Generates codecs for the datatypes in this Package
-        """
-        if codec_package_name is None:
-            codec_package_name = self._package_name + "_codecs"
-
-        if output_file_name is None:
-            codecs_path = join(self._parent.codecs_path, self._library_name)
-            file_extension = splitext(self._design_unit.source_file.name)[1]
-            output_file_name = join(codecs_path, codec_package_name + file_extension)
-
-        codec_generator.generate_codecs(
-            self._design_unit, codec_package_name, used_packages, output_file_name
-        )
-
-        return self._parent.add_source_files(output_file_name, self._library_name)
-
-
-class Test(object):
-    """
-    User interface of a single test case
-
-    """
-
-    def __init__(self, test_case):
-        self._test_case = test_case
-
-    @property
-    def name(self):
-        """
-        :returns: the entity or module name of the test bench
-        """
-        return self._test_case.name
-
-    def add_config(  # pylint: disable=too-many-arguments
-        self,
-        name,
-        generics=None,
-        parameters=None,
-        pre_config=None,
-        post_check=None,
-        sim_options=None,
-        attributes=None,
-    ):
-        """
-        Add a configuration to this test copying the default configuration.
-
-        Multiple configuration may be added one after another.
-        If no |configurations| are added the default configuration is used.
-
-        :param name: The name of the configuration. Will be added as a suffix on the test name
-        :param generics: A `dict` containing the generics to be set in addition to the default configuration.
-        :param parameters: A `dict` containing the parameters to be set in addition to the default configuration.
-        :param pre_config: A :ref:`callback function <pre_and_post_hooks>` to be called before test execution.
-        :param post_check: A :ref:`callback function <pre_and_post_hooks>` to be called after test execution.
-        :param sim_options: A `dict` containing the sim_options to be set in addition to the default configuration.
-        :param attributes: A `dict` containing the attributes to be set in addition to the default configuration.
-
-        :example:
-
-        Given the ``lib.test_bench.test`` test and the following ``add_config`` calls:
-
-        .. code-block:: python
-
-           for data_width in range(14, 15+1):
-               for sign in [False, True]:
-                   test.add_config(
-                       name="data_width=%s,sign=%s" % (data_width, sign),
-                       generics=dict(data_width=data_width, sign=sign))
-
-        The following tests will be created:
-
-        * ``lib.test_bench.data_width=14,sign=False.test``
-
-        * ``lib.test_bench.data_width=14,sign=True.test``
-
-        * ``lib.test_bench.data_width=15,sign=False.test``
-
-        * ``lib.test_bench.data_width=15,sign=True.test``
-
-        """
-        generics = {} if generics is None else generics
-        generics = lower_generics(generics)
-        parameters = {} if parameters is None else parameters
-        generics.update(parameters)
-        attributes = {} if attributes is None else attributes
-        self._test_case.add_config(
-            name=name,
-            generics=generics,
-            pre_config=pre_config,
-            post_check=post_check,
-            sim_options=sim_options,
-            attributes=attributes,
-        )
-
-    def set_attribute(self, name, value):
-        """
-        Set a value of attribute within all |configurations| of this test
-
-        :param name: The name of the attribute
-        :param value: The value of the attribute
-
-        :example:
-
-        .. code-block:: python
-
-           test.set_attribute(".foo", "bar")
-
-        """
-        self._test_case.set_attribute(name, value)
-
-    def set_generic(self, name, value):
-        """
-        Set a value of generic within all |configurations| of this test
-
-        :param name: The name of the generic
-        :param value: The value of the generic
-
-        :example:
-
-        .. code-block:: python
-
-           test.set_generic("data_width", 16)
-
-        """
-        self._test_case.set_generic(name.lower(), value)
-
-    def set_parameter(self, name, value):
-        """
-        Set a value of parameter within all |configurations| of this test
-
-        :param name: The name of the parameter
-        :param value: The value of the parameter
-
-        :example:
-
-        .. code-block:: python
-
-           test.set_parameter("data_width", 16)
-
-        """
-        self._test_case.set_generic(name, value)
-
-    def set_sim_option(self, name, value, overwrite=True):
-        """
-        Set simulation option within all |configurations| of this test
-
-        :param name: |simulation_options|
-        :param value: The value of the simulation option
-        :param overwrite: To overwrite the option or append to the existing value
-
-        :example:
-
-        .. code-block:: python
-
-           test.set_sim_option("ghdl.flags", ["--no-vital-checks"])
-
-        """
-        self._test_case.set_sim_option(name, value, overwrite)
-
-    def set_pre_config(self, value):
-        """
-        Set :ref:`pre_config <pre_and_post_hooks>` function of all |configurations| of this test
-
-        :param value: The pre_config function
-        """
-        self._test_case.set_pre_config(value)
-
-    def set_post_check(self, value):
-        """
-        Set :ref:`post_check <pre_and_post_hooks>` function of all |configurations| of this test
-
-        :param value: The post_check function
-        """
-        self._test_case.set_post_check(value)
-
-
-class SourceFileList(list):
-    """
-    A list of :class:`.SourceFile`
-    """
-
-    def __init__(self, source_files):
-        list.__init__(self, source_files)
-
-    def set_compile_option(self, name, value):
-        """
-        Set compile option for all files in the list
-
-        :param name: |compile_option|
-        :param value: The value of the compile option
-
-        :example:
-
-        .. code-block:: python
-
-           files.set_compile_option("ghdl.flags", ["--no-vital-checks"])
-        """
-        for source_file in self:
-            source_file.set_compile_option(name, value)
-
-    def add_compile_option(self, name, value):
-        """
-        Add compile option to all files in the list
-
-        :param name: |compile_option|
-        :param value: The value of the compile option
-        """
-        for source_file in self:
-            source_file.add_compile_option(name, value)
-
-    def add_dependency_on(self, source_file):
-        """
-        Add manual dependency of these files on other file(s)
-
-        :param source_file: The file(s) which this file depends on
-
-        :example:
-
-        .. code-block:: python
-
-           other_file = lib.get_source_file("other_file.vhd")
-           files.add_dependency_on(other_file)
-        """
-        for my_source_file in self:
-            my_source_file.add_dependency_on(source_file)
-
-
-class SourceFile(object):
-    """
-    A single file
-    """
-
-    def __init__(self, source_file, project, ui):
-        self._source_file = source_file
-        self._project = project
-        self._ui = ui
-
-    @property
-    def name(self):
-        """
-        The name of the SourceFile
-        """
-        return ostools.simplify_path(self._source_file.name)
-
-    @property
-    def vhdl_standard(self):
-        """
-        The VHDL standard applicable to the file,
-        None if not a VHDL file
-        """
-        if self._source_file.file_type == "vhdl":
-            return str(self._source_file.get_vhdl_standard())
-
-        return None
-
-    @property
-    def library(self):
-        """
-        The library of the source file
-        """
-        return self._ui.library(self._source_file.library.name)
-
-    def set_compile_option(self, name, value):
-        """
-        Set compile option for this file
-
-        :param name: |compile_option|
-        :param value: The value of the compile option
-
-        :example:
-
-        .. code-block:: python
-
-           my_file.set_compile_option("ghdl.flags", ["--no-vital-checks"])
-        """
-        self._source_file.set_compile_option(name, value)
-
-    def add_compile_option(self, name, value):
-        """
-        Add compile option to this file
-
-        :param name: |compile_option|
-        :param value: The value of the compile option
-        """
-        self._source_file.add_compile_option(name, value)
-
-    def get_compile_option(self, name):
-        """
-        Return compile option of this file
-
-        :param name: |compile_option|
-        """
-        return self._source_file.get_compile_option(name)
-
-    def add_dependency_on(self, source_file):
-        """
-        Add manual dependency of this file other file(s)
-
-        :param source_file: The file(s) which this file depends on
-
-        :example:
-
-        .. code-block:: python
-
-           other_files = lib.get_source_files("*.vhd")
-           my_file.add_dependency_on(other_files)
-        """
-        if isinstance(source_file, SourceFile):
-            private_source_file = (
-                source_file._source_file  # pylint: disable=protected-access
-            )
-            self._project.add_manual_dependency(
-                self._source_file, depends_on=private_source_file
-            )
-        elif hasattr(source_file, "__iter__"):
-            for element in source_file:
-                self.add_dependency_on(element)
-        else:
-            raise ValueError(source_file)
-
-
-class Results(object):
-    """
-    Gives access to results after running tests.
-    """
-
-    def __init__(self, simulator_if):
-        self._simulator_if = simulator_if
-
-    def merge_coverage(self, file_name, args=None):
-        """
-        Create a merged coverage report from the individual coverage files
-
-        :param file_name: The resulting coverage file name.
-        :param args: The tool arguments for the merge command. Should be a list of strings.
-        """
-
-        self._simulator_if.merge_coverage(file_name=file_name, args=args)
-
-
-def select_vhdl_standard(vhdl_standard=None):
-    """
-    Select VHDL standard either from class initialization or according to environment variable VUNIT_VHDL_STANDARD
-    """
-    if vhdl_standard is not None:
-        return VHDL.standard(vhdl_standard)
-
-    try:
-        return VHDL.standard(os.environ.get("VUNIT_VHDL_STANDARD", "2008"))
-    except ValueError:
-        LOGGER.error("Invalid standard set by VUNIT_VHDL_STANDARD environment variable")
-        raise
-
-
-def lower_generics(generics):
-    """
-    Convert all generics names to lower case to match internal representation.
-    @TODO Maybe warn in case of conflict. VHDL forbids this though so the user will notice anyway.
-    """
-    return dict((name.lower(), value) for name, value in generics.items())
-
-
-def check_not_empty(lst, allow_empty, error_msg):
-    """
-    Raise ValueError if the list is empty unless allow_empty is True
-    Returns the list
-    """
-    if (not allow_empty) and (not lst):
-        raise ValueError(error_msg + ". Use allow_empty=True to avoid exception.")
-    return lst
